+++ src/components/Authentication/Login/index.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import Form from "../Form/Form"
+import Form from "../Form"
 import "../Authentication.css"
 
 const Login = props => {
