+++ src/components/Game/Controls/index.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "./Button/Button";
+import Button from "./Button";
 import "nes.css/css/nes.min.css";
 import "./Controls.css";
