+++ src/components/Game/TextBox/index.js
@@ -1,14 +1,14 @@
 import React from "react";
 import "nes.css/css/nes.min.css"
+import "./TextBox.css"
 
 function TextBox(props) {
     return(
-        <div className="textbox-wrapper">
             <div className="nes-container with-title">
                 <p className="title">{props.info.title}</p>
                 <p>{props.info.description}</p>
+                {props.info.error_msg && (<p>{props.info.error_msg}</p>)}
             </div>
-        </div>
     )
 }
