+++ src/components/Game/index.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react";
-import Controls from "./Controls/Controls";
+import Controls from "./Controls";
 import TextBox from "./TextBox";
 import config from "../../config";
+import "./Game.css";
 
 export default class Game extends Component {
   state = {
@@ -33,7 +34,12 @@ export default class Game extends Component {
         return this.refresh(res.data);
       })
       .catch(err => console.log(err));
-  }
+  };
+
+  logout = () => {
+    localStorage.clear();
+    this.props.history.push("/login");
+  };
 
   componentDidMount() {
     config
@@ -45,9 +51,20 @@ export default class Game extends Component {
 
   render() {
     return (
-      <div className="nes-container">
-        <Controls move={this.movePlayer} />
-        <TextBox info={this.state} />
+      <div className="game-wrapper">
+        <div className="logout-wrapper">
+          <button className="nes-btn" id="logout-button" type="button">
+            <i class="nes-icon close" />
+          </button>
+        </div>
+        <div className="player-panel">
+          <div className="controls-wrapper">
+            <Controls move={this.movePlayer} />
+          </div>
+          <div className="textbox-wrapper">
+            <TextBox info={this.state} />
+          </div>
+        </div>
       </div>
     );
   }
